move AnnyoingTextSystem to it's own mod, removed old csproj and solutions
