+++ Mods/AnnoyingTextSystem/modinfo.json
@@ -1,0 +1,11 @@
+{
+  "type": "code",
+  "modid": "annoyingtextsystem",
+  "name": "AnnoyingTextSystem",
+  "authors": ["Tyron"],
+  "description": "",
+  "version": "1.0.0",
+  "dependency": {
+    "game": "1.12.14"
+  }
+}
