--- VSExampleMods.csproj.user
@@ -1,21 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'">
-    <StartAction>Program</StartAction>
-    <StartProgram>$(AppData)\Vintagestory\Vintagestory.exe</StartProgram>
-    <StartArguments>--addOrigin="C:\Users\tyron\Documents\vintagestory\vsmodexamples\assets" --addModPath="C:\Users\tyron\Documents\vintagestory\vsmodexamples\bin\Debug"</StartArguments>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|AnyCPU'">
-    <StartAction>Program</StartAction>
-    <StartProgram>$(AppData)\Vintagestory\Vintagestory.exe</StartProgram>
-    <StartArguments>/flatworld</StartArguments>
-  </PropertyGroup>
-  <PropertyGroup>
-    <ReferencePath>$(AppData)\Vintagestory\;$(AppData)\Vintagestory\Lib\</ReferencePath>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug +Flatworld|AnyCPU'">
-    <StartProgram>C:\Users\tyron\AppData\Roaming\Vintagestory\Vintagestory.exe</StartProgram>
-    <StartArguments>--addOrigin="C:\Users\tyron\Documents\vintagestory\vsmodexamples\assets" -osamplemods --addModPath="C:\Users\tyron\Documents\vintagestory\vsmodexamples\bin\Debug"</StartArguments>
-    <StartAction>Program</StartAction>
-  </PropertyGroup>
-</Project>
