--- VSExampleMods.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.757
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VSExampleMods", "VSExampleMods.csproj", "{203DFBF1-3599-43FD-8487-E1C79C2B788F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug + Flatworld|Any CPU = Debug + Flatworld|Any CPU
-		Debug|Any CPU = Debug|Any CPU
-		Release x64|Any CPU = Release x64|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Debug + Flatworld|Any CPU.ActiveCfg = Debug +Flatworld|Any CPU
-		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Debug + Flatworld|Any CPU.Build.0 = Debug +Flatworld|Any CPU
-		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Release x64|Any CPU.ActiveCfg = Release|Any CPU
-		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Release x64|Any CPU.Build.0 = Release|Any CPU
-		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CB2CD40B-2D5C-4630-96C4-079C5257B31C}
-	EndGlobalSection
-EndGlobal
