+++ VSMods.sln
@@ -1,0 +1,35 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30011.22
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedirectLogs", "Mods\RedirectLogs\RedirectLogs.csproj", "{203DFBF1-3599-43FD-8487-E1C79C2B788F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AnnoyingTextSystem", "Mods\AnnoyingTextSystem\AnnoyingTextSystem.csproj", "{E5887703-490B-4F5C-AAB3-BA12D0A51EB0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release x64|Any CPU = Release x64|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Release x64|Any CPU.ActiveCfg = Release|Any CPU
+		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Release x64|Any CPU.Build.0 = Release|Any CPU
+		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{203DFBF1-3599-43FD-8487-E1C79C2B788F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E5887703-490B-4F5C-AAB3-BA12D0A51EB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E5887703-490B-4F5C-AAB3-BA12D0A51EB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E5887703-490B-4F5C-AAB3-BA12D0A51EB0}.Release x64|Any CPU.ActiveCfg = Release|Any CPU
+		{E5887703-490B-4F5C-AAB3-BA12D0A51EB0}.Release x64|Any CPU.Build.0 = Release|Any CPU
+		{E5887703-490B-4F5C-AAB3-BA12D0A51EB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E5887703-490B-4F5C-AAB3-BA12D0A51EB0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {26630D08-373A-4935-AEEC-2864013A5F5B}
+	EndGlobalSection
+EndGlobal
